Add commentInteraction.js file
