+++ src/modules/app.js
@@ -1,4 +1,6 @@
-export function fetchAndRenderData() {
+import { commentModal } from './commentInteraction.js';
+
+export async function fetchAndRenderData() {
   document.addEventListener('DOMContentLoaded', () => {
     // Fetch data from the TVmaze API
     fetch('https://api.tvmaze.com/schedule?country=US&date=2014-12-01')
@@ -39,42 +41,9 @@ export function fetchAndRenderData() {
           movieDiv.appendChild(h2);
           movieDiv.appendChild(buttonsDiv);
           movieList.appendChild(movieDiv);
-
-          let modalActiveContent;
-          const commentBtns = document.querySelectorAll('.comment-button');
-          const element = document.querySelector('.modal');
-
-          // end event listerner for opening modal
-          window.closeModal = () => {
-            document.getElementById('myModal').innerHTML = '';
-            document.getElementById('myModal').style.display = 'none';
-            modalActiveContent = null;
-          };
-          // eslint-disable-next-line no-plusplus
-          for (let i = 0; i < commentBtns.length; i++) {
-            // eslint-disable-next-line no-loop-func
-            commentBtns[i].addEventListener('click', () => {
-              document.getElementById('myModal').style.display = 'block';
-              if (i === commentBtns.length - 1) {
-                modalActiveContent = data[i];
-                const template = `
-                <div class="modal-content">
-                <span onclick="closeModal()" id="close-icon" class="close">&times;</span>
-                  <img src="${modalActiveContent.show.image.medium}" id="movieImg" alt="">
-                  <h1>${modalActiveContent.show.name}</h1>
-                  <div class="details">
-                    <p>Season: ${modalActiveContent.season}</p>
-                    <p>Type: ${modalActiveContent.show.type}</p>
-                    <p>Language: ${modalActiveContent.show.language}</p>
-                    <p>Time: ${modalActiveContent.airtime}</p>
-                  </div>
-                </div>
-                `;
-                element.innerHTML += template;
-              }
-            });
-          }
+          localStorage.setItem('movies', JSON.stringify(data));
         });
+        commentModal();
       });
   });
 }
